move dataUtils
