+++ src/api/dataUtils.js
@@ -4,29 +4,29 @@ import findIndex from 'lodash/findIndex'
 import filter from 'lodash/filter'
 import sortBy from 'lodash/sortBy'
 
-import airlinguist from './imgs/AirLinguist.png'
-import airudite from './imgs/Airudite.png'
-import arslonga from './imgs/ArsLonga.png'
-import dystopiair from './imgs/DystopiAir.png'
-import gobelow from './imgs/GoBelow.png'
-import gossamerica from './imgs/GossAmerica.png'
-import incorporeal from './imgs/IncorpoREAL.png'
-import janeair from './imgs/JaneAir.png'
-import oddyssey from './imgs/Oddyssey.png'
-import panache from './imgs/Panache.png'
-import quantum from './imgs/Quantum.png'
-import spellbound from './imgs/SpellboundAirlines.png'
-import untitledairlines from './imgs/UntitledAirlines.png'
-import utopiair from './imgs/UtopiAir.png'
-import wistful from './imgs/Wistful.png'
+import airlinguist from '../imgs/AirLinguist.png'
+import airudite from '../imgs/Airudite.png'
+import arslonga from '../imgs/ArsLonga.png'
+import dystopiair from '../imgs/DystopiAir.png'
+import gobelow from '../imgs/GoBelow.png'
+import gossamerica from '../imgs/GossAmerica.png'
+import incorporeal from '../imgs/IncorpoREAL.png'
+import janeair from '../imgs/JaneAir.png'
+import oddyssey from '../imgs/Oddyssey.png'
+import panache from '../imgs/Panache.png'
+import quantum from '../imgs/Quantum.png'
+import spellbound from '../imgs/SpellboundAirlines.png'
+import untitledairlines from '../imgs/UntitledAirlines.png'
+import utopiair from '../imgs/UtopiAir.png'
+import wistful from '../imgs/Wistful.png'
 
 // audio files
-import abame from './sound/announcement-abame.mp3'
-import ankhmorpark from './sound/announcement-ankhmorpark.mp3'
-import asteroidb612 from './sound/announcement-asteroid612.mp3'
-import cimeria from './sound/announcement-cimeria.mp3'
-import cityofbrass from './sound/announcement-cityofbrass.mp3'
-import cyberspace from './sound/announcement-cyberspace.mp3'
+import abame from '../sound/announcement-abame.mp3'
+import ankhmorpark from '../sound/announcement-ankhmorpark.mp3'
+import asteroidb612 from '../sound/announcement-asteroid612.mp3'
+import cimeria from '../sound/announcement-cimeria.mp3'
+import cityofbrass from '../sound/announcement-cityofbrass.mp3'
+import cyberspace from '../sound/announcement-cyberspace.mp3'
 
 const imagepaths = {
     airlinguist,
