+++ src/api/tests/FlightsActionsTest.test.js
@@ -8,7 +8,7 @@ import configureMockStore from 'redux-mock-store'
 import thunk from 'redux-thunk'
 import fetchMock from 'fetch-mock'
 import moment from 'moment'
-import { normalizeFlight } from '../../dataUtils'
+import { normalizeFlight } from '../dataUtils'
 
 const mockFlights = [
     {
